Add files via upload
This is what I have from what Dr. Booth gave us in class. I'm not sure if it will compile, since I wasn't able to figure out how to make the plotter work on my device, but I'm pretty sure it is the same as what he has shown in class. It includes both the tile object that can fall and move side to side, as well as the stuff on fonts that he talked about yesterday.